Added new password checker hook and updated change password page with pwd checker.
